requirements.txt
python-dotenv
sshtunnel
psycopg2-binary     # PostgreSQL adapter for Python
pandas
datetime
